+++ src/States/MapViewStateUi.cpp
@@ -96,6 +96,9 @@ void MapViewState::initUi()
 	mPopulationPanel.morale(&mCurrentMorale);
 	mPopulationPanel.old_morale(&mPreviousMorale);
 
+	mResourceBreakdownPanel.position(0, 22);
+	mResourceBreakdownPanel.playerResources(&mPlayerResources);
+
 	mGameOverDialog.returnToMainMenu().connect(this, &MapViewState::btnGameOverClicked);
 	mGameOverDialog.hide();